move headers under snowflake folder
