+++ libsnowflakeclient/examples/selectnull.c
@@ -5,9 +5,9 @@
 
 #include <stdio.h>
 #include <string.h>
-#include <snowflake_client.h>
-#include <example_setup.h>
 #include <stdlib.h>
+#include <snowflake/client.h>
+#include <example_setup.h>
 
 typedef struct test_case_to_string {
     const int64 c1in;
