+++ libsnowflakeclient/examples/timezone.c
@@ -4,9 +4,9 @@
 
 
 #include <stdio.h>
-#include <snowflake_client.h>
-#include <example_setup.h>
 #include <memory.h>
+#include <snowflake/client.h>
+#include <example_setup.h>
 
 
 int main() {
