+++ libsnowflakeclient/include/snowflake/logger.h
