+++ libsnowflakeclient/include/snowflake/version.h
