+++ libsnowflakeclient/lib/basic_types.c
@@ -2,7 +2,7 @@
  * Copyright (c) 2017 Snowflake Computing, Inc. All rights reserved.
  */
 
-#include <snowflake_basic_types.h>
+#include <snowflake/basic_types.h>
 
 const int8 SF_BOOLEAN_TRUE = 1;
 const int8 SF_BOOLEAN_FALSE = 0;
