+++ libsnowflakeclient/lib/chunk_downloader.h
@@ -17,7 +17,7 @@ extern "C" {
 
 #include <pthread.h>
 #include <curl/curl.h>
-#include <snowflake_client.h>
+#include <snowflake/client.h>
 #include "cJSON.h"
 
 typedef struct sf_queue_item {
