+++ libsnowflakeclient/lib/client_int.h
