+++ libsnowflakeclient/lib/memory.h
@@ -16,7 +16,7 @@ extern "C" {
 #endif
 
 #include <stdlib.h>
-#include <snowflake_basic_types.h>
+#include <snowflake/basic_types.h>
 
 #define SF_MALLOC(s) sf_malloc(s, __FILE__, __LINE__)
 #define SF_CALLOC(n, s) sf_calloc(n, s, __FILE__, __LINE__)
